backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
google-auth==2.40.3
google-genai==1.20.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
nltk==3.9.1
py-readability-metrics==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
textrazor==1.4.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
